ajuste frontEnd

+++ BackEnd/BackEnd.Cervejaria/SofwareContext/Cervejaria.Domain/Enuns/TipoUsuario.cs
@@ -2,7 +2,7 @@
 
 namespace Cervejaria.Domain.Enuns
 {
-    public enum TipoUsuario : Int16
+    public enum TipoUsuario : Int32
     {
         Cliente,
         Fornecedor
